Add OD::Prometheus::Query class
This class is used for talking to Prometheus servers via the HTTP
API. Implementations for understanding query and query_range queries.
We had to modify the OD::Prometheus::Metric class, in order to be
more efficient when storing the results from range queries. Now,
the value of a metric can also be an arrayref, in addition to a scalar
value which was the expected type so far. The arrayref contains
arrayrefs with two elements, the first being the timestamp and the
second the scalar value. Multiple values can thus be taken from
the result of a query_range and inserted into an OD::Prometheus::Metric
efficiently.