Rework values and value
The attribute values is kept and is now used to store even a single
value. Likewise for the timestamp attribute. Methods returning the
value and timestamp are provided for backwards compatibility. It is
still illegal to call value on multi-valued metrics. But it is no
longer illegal to call values on single-values metrics.

Method size is added, returns number of values.

Method first is added, returns first element of values

Minor modifications to other methods to accommodate above mentioned
changes